Rename models/config -> models/client
